refactor: move AI provider db-to-SDK converter into db2sdk
dbAIProviderToSDK and its decode/redact helpers now live in db2sdk as
AIProvider, AIProviderSettings, and an internal redactor. The handler
file still owns write-side encoding and the merge helper since those
are call-side concerns.
